[refactor] map,log가 사이드바를 함께 쓰도록 구조 변경
